Colocate desc str gen to the create route
