+++ src/app/api/spotify/create/_lib/requests.ts
@@ -13,7 +13,54 @@ import {
 	MyPlaylistObject,
 	SpotAlbumTracksResponse,
 	SpotPlaylistTracksResponse,
+	ActionType,
 } from '@/components/spotify/types';
+
+type descArgs = { du?: number; shared?: number; tu?: number };
+
+export function SERVER_DIFF_TYPES(params: {
+	target: {
+		name: string;
+		owner: string;
+	};
+	differ: {
+		name: string;
+		owner: string;
+	};
+	actionType: ActionType;
+}) {
+	const { target, differ, actionType } = params;
+	switch (actionType) {
+		case 'adu':
+			return ({ du }: descArgs) =>
+				`Added ${du ?? 'some'} tracks from ${differ.owner}'s "${
+					differ.name
+				}" to ` + `${target.owner} 's "${target.name}".`;
+
+		case 'odu':
+			return ({ du }: descArgs) =>
+				`Replaced ${target.owner}'s "${target.name}" entirely with ` +
+				`${du} tracks unique to ${differ.owner}'s "${differ.name}".`;
+
+		case 'otu':
+			return ({ shared }: descArgs) =>
+				`${shared ?? 'Some'} tracks shared by ${target.owner}'s "${
+					target.name
+				}" and ` + `${differ.owner}'s "${differ.name}" were removed.`;
+
+		case 'bu':
+			return ({ du, tu }: descArgs) =>
+				`Any similarities between ${target.owner}'s "${target.name}" ` +
+				`and ${differ.owner}'s "${differ.name}" were removed; ` +
+				`now contains ${tu ?? 'some'} tracks from the former ` +
+				`and ${du ?? 'some'} tracks from the latter.`;
+
+		case 'stu':
+			return ({ shared }: descArgs) =>
+				`${shared} tracks that exist in both ${target.owner}'s ` +
+				`"${target.name}" and ${differ.owner}'s "${differ.name}" remain.`;
+	}
+}
 
 function genUId(length: number) {
 	const chars = 'abcdefghijklmnopqrstuvwxyz0123456789';
